Add an example for full disk encryption
Simulate the functionality of the Android Full
Disk Encryption (FDE) by using a symmetric key to
encrypt the contents of a supplied passphrase from
the user. Save the encrypted passphrase to a file
to simulate encrypting the disk. Keep the key
encrypted in the file metadata and on the key
update, just update the encrypted key and not the
rest of the file contents

Signed-off-by: Sebastian Ene <sebastianene@google.com>